+++ .github/workflows/gdsc-dju-web-deploy-preview.yml
@@ -27,6 +27,7 @@ jobs:
         run: yarn install
       - name: Project Build
         run: yarn build
+        working-directory: ${{ env.working-directory }}
       - name: Get short SHA
         id: slug
         run: echo "::set-output name=sha8::$(echo ${GITHUB_SHA} | cut -c1-8)"
